requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
python-multipart
httpx
aiofiles
opencv-python-headless
numpy
